// src/app/model/cliente.model.ts
export class ClienteModel {
    id: string;
    Marca: number;
    razonsocial: string;
    Rut: string;
    Direccion: string;
    dir: string;
    contacto: string;
    email: string;
    Tel: string;
    Token: string;
    status: number;
    Locales: any;

    constructor() {
        this.status = 1;
    }
}